+++ tests/example_targets/c_examples/multifile_project/add_sub.c
@@ -1,9 +1,0 @@
-#include "add_sub.h"
-
-my_double Add(my_double a, my_double b) {
-    return a + b;
-}
-
-double Sub(double a, double b) {
-    return a - b;
-}