+++ tests/example_targets/c_examples/multifile_project/headers/types.h
@@ -1,2 +1,0 @@
-typedef double my_double;
-typedef float my_float;