+++ CHANGELOG.md
@@ -8,3 +8,7 @@ and this project adheres to [Semantic Versioning](http://semver.org/).
 ## 20220716
 - Add modules preprocessor to Futag python-package
 - Fix README of Futag python-package
+
+## 20220727
+- Add custom-llvm: download and build llvm, clang, compiler-rt
+- Fix document